Initial patch to support replication slots
This patch introduces 3 new functions:
* _clean_replication_slots
* _create_replication_slots
* _init_replication_slots

The first one is called for each demote operation. Maybe we should also call
it when we start PostgreSQL in standby mode.
The second and third ones are called just before pgsqlms promotes a standby
server to master.

This is just a test, it's not ready for production.